ADM-434 refactor VelocityCalculator
